Updated 3D and vulkan bugs Changes committed: 	modified:   Cargo.toml 	modified:   src/camera/perspective_camera.rs 	modified:   src/core.rs 	modified:   src/drawcalls.rs 	modified:   src/graphics.rs 	modified:   src/math.rs 	modified:   src/shaders/glsl/VkModel.frag 	modified:   src/shaders/glsl/VkModelDeffered.frag 	modified:   src/shaders/glsl/VkText.frag 	new file:   src/shaders/glsl/VkTextInstanced.vert 	modified:   src/shaders/model_shader.rs 	modified:   src/shaders/sprv/VkModelDefferedFrag.spv 	modified:   src/shaders/sprv/VkModelFrag.spv 	modified:   src/shaders/sprv/VkTextFrag.spv 	new file:   src/shaders/sprv/VkTextInstancedVert.spv 	modified:   src/shaders/texture_shader.rs

+++ src/camera/perspective_camera.rs
@@ -422,7 +422,7 @@ impl PerspectiveCamera {
     let near = 0.1;
     let far = 256.0;
     let fov = 60.0;
-    let f = 1.0 / (math::to_radians(fov) / 2.0).tan();
+    let f = 1.0 / (math::to_radians(fov as f64) as f32 / 2.0).tan();
     
     let perspective = Matrix4::from_cols(
                       Vector4::new(f / aspect, 0.0,   0.0,                               0.0),
